continuing with refactor
